+++ examples/kitchen_sink/routes/messages/index.dart
@@ -1,0 +1,6 @@
+import 'package:dart_frog/dart_frog.dart';
+
+Future<Response> onRequest(RequestContext context) async {
+  final body = await context.request.body();
+  return Response(body: 'message: $body');
+}