+++ examples/web_socket_counter/lib/counter/counter.dart
@@ -1,0 +1,3 @@
+export 'cubit/counter_cubit.dart';
+export 'middleware/counter_provider.dart';
+export 'models/message.dart';