+++ packages/dart_frog/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.3.2
+
+- feat: cache `Request` and `Response` body
+
+# 0.3.1
+
+- feat: add `formData` to `Request`/`Response`
+
 # 0.3.0
 
 - **BREAKING** fix: `Request.json()` and `Response.json()` return `Future<dynamic>`